requirements.txt
appnope==0.1.0
autopep8==1.4.3
backcall==0.1.0
confusable-homoglyphs==3.2.0
decorator==4.4.0
Django==2.2
django-redis==4.10.0
django-registration==3.0
django-sass-processor==0.7.3
djangorestframework==3.9.2
factory-boy==2.11.1
Faker==1.0.4
gunicorn==19.9.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
libsass==0.18.0
numpy==1.16.2
pandas==0.24.2
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.1
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.2
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
